implement config flow for real power sensor (#1788)
* implement config flow for real power sensor

* Add translations

* Add documentation

* Add dutch translations

* Move config flow test to separate package and files

* Add options flow
