+++ custom_components/powercalc/config_flow.py
@@ -91,6 +91,7 @@ SENSOR_TYPE_MENU = {
     SensorType.GROUP: "Group",
     SensorType.VIRTUAL_POWER: "Virtual power (manual)",
     MENU_OPTION_LIBRARY: "Virtual power (library)",
+    SensorType.REAL_POWER: "Energy from real power sensor",
 }
 
 SCHEMA_DAILY_ENERGY_OPTIONS = vol.Schema(
@@ -124,6 +125,24 @@ SCHEMA_DAILY_ENERGY = vol.Schema(
         vol.Optional(CONF_UNIQUE_ID): selector.TextSelector(),
     },
 ).extend(SCHEMA_DAILY_ENERGY_OPTIONS.schema)
+
+SCHEMA_REAL_POWER_OPTIONS = vol.Schema(
+    {
+        vol.Required(CONF_ENTITY_ID): selector.EntitySelector(
+            selector.EntitySelectorConfig(device_class=SensorDeviceClass.POWER),
+        ),
+        vol.Optional(
+            CONF_CREATE_UTILITY_METERS,
+            default=False,
+        ): selector.BooleanSelector(),
+    },
+)
+
+SCHEMA_REAL_POWER = vol.Schema(
+    {
+        vol.Required(CONF_NAME): selector.TextSelector(),
+    },
+).extend(SCHEMA_REAL_POWER_OPTIONS.schema)
 
 SCHEMA_POWER_LIBRARY = vol.Schema(
     {
@@ -507,6 +526,22 @@ class ConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
 
         return await self.async_step_manufacturer()
 
+    async def async_step_real_power(self, user_input: dict[str, Any] | None = None) -> FlowResult:
+        """Handle the flow for real power sensor"""
+
+        self.selected_sensor_type = SensorType.REAL_POWER
+        if user_input is not None:
+            self.name = user_input.get(CONF_NAME)
+            self.sensor_config.update(user_input)
+            return self.create_config_entry()
+
+        return self.async_show_form(
+            step_id="real_power",
+            data_schema=SCHEMA_REAL_POWER,
+            errors={},
+            last_step=False,
+        )
+
     async def async_step_manufacturer(
         self,
         user_input: dict[str, Any] | None = None,
@@ -745,7 +780,7 @@ class OptionsFlowHandler(OptionsFlow):
                 except StrategyConfigurationError as error:
                     return {"base": error.get_config_flow_translate_key()}
 
-        if self.sensor_type == SensorType.GROUP:
+        if self.sensor_type in [SensorType.GROUP, SensorType.REAL_POWER]:
             self._process_user_input(user_input, schema)
 
         self.hass.config_entries.async_update_entry(
@@ -796,6 +831,9 @@ class OptionsFlowHandler(OptionsFlow):
             )
 
             strategy_options = self.current_config.get(self.strategy) or {}
+
+        if self.sensor_type == SensorType.REAL_POWER:
+            data_schema = SCHEMA_REAL_POWER_OPTIONS
 
         if self.sensor_type == SensorType.DAILY_ENERGY:
             data_schema = SCHEMA_DAILY_ENERGY_OPTIONS
