+++ docs/source/real-power-sensor.rst
@@ -2,7 +2,26 @@
 Real power sensor
 =================
 
-In the yaml configuration (functionality not available through the webUI) you can add the following configuration
+Powercalc also provides possibilities to create energy sensors (and optionally utility meters) for existing power sensors in your installation.
+You could also create them with the built-in helpers HA provides, but Powercalc makes it even more easy and you can also includes existing power sensors in Powercalc groups this way.
+The energy sensors can be added to the energy dashboard.
+
+You can create this either with the YAML or GUI.
+
+GUI
+---
+
+Just click this button to start the configuration flow:
+
+.. image:: https://my.home-assistant.io/badges/config_flow_start.svg
+   :target: https://my.home-assistant.io/redirect/config_flow_start/?domain=powercalc
+
+Select :guilabel:`Energy from real power sensor` and follow the instructions.
+
+YAML
+----
+
+In the yaml configuration you can add the following configuration
 to use an existing power sensor and let powercalc create the energy sensors and utility meters for it:
 
 .. code-block:: yaml
@@ -14,8 +33,7 @@ to use an existing power sensor and let powercalc create the energy sensors and 
           force_energy_sensor_creation: true # optional
 
 This also enables you to combine virtual power sensors (created with powercalc) and existing power sensors in your HA installation into
-a group. Without this configuration option power_sensor_id that would not be possible.
+a YAML group. Without this configuration option power_sensor_id that would not be possible.
 
-If you don't define `force_energy_sensor_creation` or you set it to `false` an energy sensor will not be created if the device already
-has an energy sensor. This can be a problem if you want to create an energy sensor for an MQTT device with multiple energy and power
-sensors already in it.
+.. note::
+    If you don't define `force_energy_sensor_creation` or you set it to `false` an energy sensor will not be created if the device already has an energy sensor. This can be a problem if you want to create an energy sensor for an MQTT device with multiple energy and power sensors already in it.
