--- tests/test_config_flow.py
@@ -1,1318 +1,0 @@
-from collections.abc import Mapping
-from typing import Any
-from unittest.mock import patch
-
-import pytest
-import voluptuous as vol
-from homeassistant import config_entries, data_entry_flow
-from homeassistant.components import sensor
-from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import (
-    CONF_ENTITIES,
-    CONF_ENTITY_ID,
-    CONF_NAME,
-    CONF_PLATFORM,
-    CONF_UNIQUE_ID,
-    CONF_UNIT_OF_MEASUREMENT,
-    STATE_ON,
-    UnitOfPower,
-)
-from homeassistant.core import HomeAssistant
-from homeassistant.data_entry_flow import FlowResult
-from homeassistant.helpers.area_registry import AreaRegistry
-from homeassistant.helpers.entity_registry import EntityRegistry
-from homeassistant.helpers.selector import SelectSelector
-from homeassistant.helpers.typing import ConfigType
-from homeassistant.setup import async_setup_component
-
-import custom_components.test.sensor as test_sensor_platform
-from custom_components.powercalc.common import create_source_entity
-from custom_components.powercalc.config_flow import (
-    CONF_CONFIRM_AUTODISCOVERED_MODEL,
-    DOMAIN,
-    MENU_OPTION_LIBRARY,
-)
-from custom_components.powercalc.const import (
-    CONF_AREA,
-    CONF_CALCULATION_ENABLED_CONDITION,
-    CONF_CREATE_ENERGY_SENSOR,
-    CONF_CREATE_UTILITY_METERS,
-    CONF_DAILY_FIXED_ENERGY,
-    CONF_ENERGY_INTEGRATION_METHOD,
-    CONF_FIXED,
-    CONF_GROUP_MEMBER_SENSORS,
-    CONF_GROUP_POWER_ENTITIES,
-    CONF_HIDE_MEMBERS,
-    CONF_LINEAR,
-    CONF_MANUFACTURER,
-    CONF_MAX_POWER,
-    CONF_MIN_POWER,
-    CONF_MODE,
-    CONF_MODEL,
-    CONF_POWER,
-    CONF_POWER_FACTOR,
-    CONF_POWER_TEMPLATE,
-    CONF_SENSOR_TYPE,
-    CONF_STATES_POWER,
-    CONF_SUB_GROUPS,
-    CONF_SUB_PROFILE,
-    CONF_UPDATE_FREQUENCY,
-    CONF_VALUE,
-    CONF_VOLTAGE,
-    CONF_WLED,
-    DISCOVERY_POWER_PROFILE,
-    DISCOVERY_SOURCE_ENTITY,
-    ENERGY_INTEGRATION_METHOD_LEFT,
-    CalculationStrategy,
-    SensorType,
-)
-from custom_components.powercalc.discovery import ModelInfo, autodiscover_model
-from custom_components.powercalc.errors import StrategyConfigurationError
-from custom_components.powercalc.power_profile.factory import get_power_profile
-from custom_components.test.light import MockLight
-
-from .common import (
-    MockConfigEntry,
-    create_mock_light_entity,
-    create_mocked_virtual_power_sensor_entry,
-)
-from .conftest import MockEntityWithModel
-
-DEFAULT_ENTITY_ID = "light.test"
-DEFAULT_UNIQUE_ID = "7c009ef6829f"
-
-
-async def test_discovery_flow(
-    hass: HomeAssistant,
-    mock_entity_with_model_information: MockEntityWithModel,
-) -> None:
-    mock_entity_with_model_information(
-        "light.test",
-        "signify",
-        "LCT010",
-        unique_id=DEFAULT_UNIQUE_ID,
-    )
-
-    source_entity = await create_source_entity(DEFAULT_ENTITY_ID, hass)
-    power_profile = await get_power_profile(
-        hass,
-        {},
-        await autodiscover_model(hass, source_entity.entity_entry),
-    )
-
-    result: FlowResult = await hass.config_entries.flow.async_init(
-        DOMAIN,
-        context={"source": config_entries.SOURCE_INTEGRATION_DISCOVERY},
-        data={
-            CONF_UNIQUE_ID: DEFAULT_UNIQUE_ID,
-            CONF_NAME: "test",
-            CONF_ENTITY_ID: DEFAULT_ENTITY_ID,
-            CONF_MANUFACTURER: "signify",
-            CONF_MODEL: "LCT010",
-            DISCOVERY_SOURCE_ENTITY: source_entity,
-            DISCOVERY_POWER_PROFILE: power_profile,
-        },
-    )
-
-    # Confirm selected manufacturer/model
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        {CONF_CONFIRM_AUTODISCOVERED_MODEL: True},
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    assert result["data"] == {
-        CONF_ENTITY_ID: DEFAULT_ENTITY_ID,
-        CONF_SENSOR_TYPE: SensorType.VIRTUAL_POWER,
-        CONF_MANUFACTURER: "signify",
-        CONF_MODEL: "LCT010",
-        CONF_NAME: "test",
-        CONF_UNIQUE_ID: f"pc_{DEFAULT_UNIQUE_ID}",
-    }
-
-
-async def test_discovery_flow_remarks_are_shown(hass: HomeAssistant) -> None:
-    """Model.json can provide remarks to show in the discovery flow. Check if these are displayed correctly"""
-    source_entity = await create_source_entity("media_player.test", hass)
-    power_profile = await get_power_profile(hass, {}, ModelInfo("sonos", "one"))
-
-    result: FlowResult = await hass.config_entries.flow.async_init(
-        DOMAIN,
-        context={"source": config_entries.SOURCE_INTEGRATION_DISCOVERY},
-        data={
-            CONF_UNIQUE_ID: DEFAULT_UNIQUE_ID,
-            CONF_NAME: "test",
-            CONF_ENTITY_ID: "media_player.test",
-            CONF_MANUFACTURER: "sonos",
-            CONF_MODEL: "one",
-            DISCOVERY_SOURCE_ENTITY: source_entity,
-            DISCOVERY_POWER_PROFILE: power_profile,
-        },
-    )
-    assert result["description_placeholders"]["remarks"] is not None
-
-
-async def test_discovery_flow_with_subprofile_selection(
-    hass: HomeAssistant,
-    mock_entity_with_model_information: MockEntityWithModel,
-) -> None:
-    mock_entity_with_model_information(
-        "light.test",
-        "lifx",
-        "LIFX Z",
-        unique_id=DEFAULT_UNIQUE_ID,
-    )
-
-    source_entity = await create_source_entity(DEFAULT_ENTITY_ID, hass)
-    power_profile = await get_power_profile(
-        hass,
-        {},
-        await autodiscover_model(hass, source_entity.entity_entry),
-    )
-
-    result: FlowResult = await hass.config_entries.flow.async_init(
-        DOMAIN,
-        context={"source": config_entries.SOURCE_INTEGRATION_DISCOVERY},
-        data={
-            CONF_ENTITY_ID: DEFAULT_ENTITY_ID,
-            CONF_NAME: "test",
-            CONF_MANUFACTURER: "lifx",
-            CONF_MODEL: "LIFX Z",
-            CONF_UNIQUE_ID: DEFAULT_UNIQUE_ID,
-            DISCOVERY_SOURCE_ENTITY: source_entity,
-            DISCOVERY_POWER_PROFILE: power_profile,
-        },
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        {CONF_CONFIRM_AUTODISCOVERED_MODEL: True},
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["step_id"] == "sub_profile"
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        {CONF_SUB_PROFILE: "length_6"},
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    assert result["data"] == {
-        CONF_ENTITY_ID: DEFAULT_ENTITY_ID,
-        CONF_SENSOR_TYPE: SensorType.VIRTUAL_POWER,
-        CONF_MANUFACTURER: "lifx",
-        CONF_MODEL: "LIFX Z/length_6",
-        CONF_NAME: "test",
-        CONF_UNIQUE_ID: f"pc_{DEFAULT_UNIQUE_ID}",
-    }
-
-
-async def test_sensor_type_menu_displayed(hass: HomeAssistant) -> None:
-    """Test a menu is diplayed with sensor type selection"""
-
-    result: FlowResult = await hass.config_entries.flow.async_init(
-        DOMAIN,
-        context={"source": config_entries.SOURCE_USER},
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.MENU
-    assert result["step_id"] == "user"
-
-
-@pytest.mark.parametrize(
-    "sensor_type",
-    [SensorType.VIRTUAL_POWER, SensorType.DAILY_ENERGY, SensorType.GROUP],
-)
-async def test_sensor_type_form_displayed(
-    hass: HomeAssistant,
-    sensor_type: SensorType,
-) -> None:
-    await _select_sensor_type(hass, sensor_type)
-
-
-async def test_create_fixed_sensor_entry(hass: HomeAssistant) -> None:
-    result = await _goto_virtual_power_strategy_step(hass, CalculationStrategy.FIXED)
-    result = await _set_virtual_power_configuration(hass, result, {CONF_POWER: 20})
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    _assert_default_virtual_power_entry_data(
-        CalculationStrategy.FIXED,
-        result["data"],
-        {CONF_FIXED: {CONF_POWER: 20}},
-    )
-
-    await hass.async_block_till_done()
-    assert hass.states.get("sensor.test_power")
-    assert hass.states.get("sensor.test_energy")
-
-
-async def test_create_fixed_sensor_entry_with_template(hass: HomeAssistant) -> None:
-    result = await _goto_virtual_power_strategy_step(hass, CalculationStrategy.FIXED)
-
-    template = "{{states(input_number.my_number)}} | float"
-    result = await _set_virtual_power_configuration(
-        hass,
-        result,
-        {CONF_POWER_TEMPLATE: template},
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    _assert_default_virtual_power_entry_data(
-        CalculationStrategy.FIXED,
-        result["data"],
-        {CONF_FIXED: {CONF_POWER_TEMPLATE: template}},
-    )
-
-    await hass.async_block_till_done()
-    assert hass.states.get("sensor.test_power")
-    assert hass.states.get("sensor.test_energy")
-
-
-async def test_create_fixed_sensor_entry_with_states_power(hass: HomeAssistant) -> None:
-    result = await _goto_virtual_power_strategy_step(hass, CalculationStrategy.FIXED)
-    result = await _set_virtual_power_configuration(
-        hass,
-        result,
-        {CONF_STATES_POWER: {"playing": 1.8}},
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    _assert_default_virtual_power_entry_data(
-        CalculationStrategy.FIXED,
-        result["data"],
-        {CONF_FIXED: {CONF_STATES_POWER: {"playing": 1.8}}},
-    )
-
-    await hass.async_block_till_done()
-    assert hass.states.get("sensor.test_power")
-    assert hass.states.get("sensor.test_energy")
-
-
-async def test_create_linear_sensor_entry(hass: HomeAssistant) -> None:
-    result = await _goto_virtual_power_strategy_step(hass, CalculationStrategy.LINEAR)
-    result = await _set_virtual_power_configuration(
-        hass,
-        result,
-        {CONF_MIN_POWER: 1, CONF_MAX_POWER: 40},
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    _assert_default_virtual_power_entry_data(
-        CalculationStrategy.LINEAR,
-        result["data"],
-        {CONF_LINEAR: {CONF_MIN_POWER: 1, CONF_MAX_POWER: 40}},
-    )
-
-    await hass.async_block_till_done()
-    assert hass.states.get("sensor.test_power")
-    assert hass.states.get("sensor.test_energy")
-
-
-async def test_create_linear_sensor_error_mandatory_fields(hass: HomeAssistant) -> None:
-    result = await _goto_virtual_power_strategy_step(hass, CalculationStrategy.LINEAR)
-    result = await _set_virtual_power_configuration(hass, result, {CONF_MIN_POWER: 20})
-
-    assert result["errors"]
-    assert result["errors"]["base"] == "linear_mandatory"
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-
-
-async def test_create_wled_sensor_entry(hass: HomeAssistant) -> None:
-    await _create_wled_entities(hass)
-
-    result = await _goto_virtual_power_strategy_step(hass, CalculationStrategy.WLED)
-    result = await _set_virtual_power_configuration(
-        hass,
-        result,
-        {CONF_VOLTAGE: 12, CONF_POWER_FACTOR: 0.8},
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    _assert_default_virtual_power_entry_data(
-        CalculationStrategy.WLED,
-        result["data"],
-        {CONF_WLED: {CONF_VOLTAGE: 12, CONF_POWER_FACTOR: 0.8}},
-    )
-
-    await hass.async_block_till_done()
-    assert hass.states.get("sensor.test_power")
-    assert hass.states.get("sensor.test_energy")
-
-
-async def test_lut_manual_flow(hass: HomeAssistant) -> None:
-    light_entity = MockLight("test", STATE_ON, DEFAULT_UNIQUE_ID)
-    await create_mock_light_entity(hass, light_entity)
-
-    result = await _goto_virtual_power_strategy_step(hass, CalculationStrategy.LUT)
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["step_id"] == "manufacturer"
-    data_schema: vol.Schema = result["data_schema"]
-    manufacturer_select: SelectSelector = data_schema.schema["manufacturer"]
-    manufacturer_options = manufacturer_select.config["options"]
-    assert {"value": "belkin", "label": "belkin"} in manufacturer_options
-    assert {"value": "signify", "label": "signify"} in manufacturer_options
-
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        {CONF_MANUFACTURER: "signify"},
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["step_id"] == "model"
-    data_schema: vol.Schema = result["data_schema"]
-    model_select: SelectSelector = data_schema.schema["model"]
-    model_options = model_select.config["options"]
-    assert {"value": "LCT010", "label": "LCT010"} in model_options
-    assert {"value": "LWB010", "label": "LWB010"} in model_options
-
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        {CONF_MODEL: "LCT010"},
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-
-    # Advanced options step
-    result = await hass.config_entries.flow.async_configure(result["flow_id"], {})
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-
-    _assert_default_virtual_power_entry_data(
-        CalculationStrategy.LUT,
-        result["data"],
-        {CONF_MANUFACTURER: "signify", CONF_MODEL: "LCT010"},
-    )
-
-    await hass.async_block_till_done()
-    assert hass.states.get("sensor.test_power")
-    assert hass.states.get("sensor.test_energy")
-
-
-async def test_lut_autodiscover_flow(
-    hass: HomeAssistant,
-    mock_entity_with_model_information: MockEntityWithModel,
-) -> None:
-    manufacturer = "ikea"
-    model = "LED1545G12"
-    mock_entity_with_model_information(
-        "light.test",
-        manufacturer,
-        model,
-        unique_id=DEFAULT_UNIQUE_ID,
-    )
-
-    result = await _goto_virtual_power_strategy_step(hass, CalculationStrategy.LUT)
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["step_id"] == "library"
-    assert result["description_placeholders"] == {
-        "manufacturer": "ikea",
-        "model": "LED1545G12",
-        "remarks": None,
-    }
-
-    result = await _set_virtual_power_configuration(
-        hass,
-        result,
-        {CONF_CONFIRM_AUTODISCOVERED_MODEL: True},
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    _assert_default_virtual_power_entry_data(
-        CalculationStrategy.LUT,
-        result["data"],
-        {CONF_MANUFACTURER: manufacturer, CONF_MODEL: model},
-    )
-
-    await hass.async_block_till_done()
-    assert hass.states.get("sensor.test_power")
-    assert hass.states.get("sensor.test_energy")
-
-
-async def test_lut_not_autodiscovered_model_unsupported(
-    hass: HomeAssistant,
-    mock_entity_with_model_information: MockEntityWithModel,
-) -> None:
-    mock_entity_with_model_information("light.test", "ikea", "unknown_model")
-
-    result = await _goto_virtual_power_strategy_step(hass, CalculationStrategy.LUT)
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["step_id"] == "manufacturer"
-
-
-async def test_lut_not_autodiscovered(hass: HomeAssistant) -> None:
-    light_entity = MockLight("test", STATE_ON)
-    light_entity._attr_unique_id = None  # noqa
-    await create_mock_light_entity(hass, light_entity)
-
-    result = await _goto_virtual_power_strategy_step(hass, CalculationStrategy.LUT)
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["step_id"] == "manufacturer"
-
-
-async def test_lut_autodiscover_flow_not_confirmed(
-    hass: HomeAssistant,
-    mock_entity_with_model_information: MockEntityWithModel,
-) -> None:
-    """
-    When manufacturer and model are auto detected and user chooses to not accept it,
-    make sure he/she is forwarded to the manufacturer listing
-    """
-    mock_entity_with_model_information(
-        "light.test",
-        "ikea",
-        "LED1545G12",
-        unique_id="234438",
-    )
-
-    result = await _goto_virtual_power_strategy_step(hass, CalculationStrategy.LUT)
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["step_id"] == "library"
-
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        {CONF_CONFIRM_AUTODISCOVERED_MODEL: False},
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["step_id"] == "manufacturer"
-
-
-async def test_lut_flow_with_sub_profiles(
-    hass: HomeAssistant,
-    mock_entity_with_model_information: MockEntityWithModel,
-) -> None:
-    mock_entity_with_model_information("light.test", "", "")
-
-    result = await _goto_virtual_power_strategy_step(hass, CalculationStrategy.LUT)
-
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        {CONF_MANUFACTURER: "yeelight"},
-    )
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        {CONF_MODEL: "YLDL01YL"},
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["step_id"] == "sub_profile"
-    data_schema: vol.Schema = result["data_schema"]
-    model_select: SelectSelector = data_schema.schema["sub_profile"]
-    select_options = model_select.config["options"]
-    assert {"value": "ambilight", "label": "ambilight"} in select_options
-    assert {"value": "downlight", "label": "downlight"} in select_options
-
-    result = await _set_virtual_power_configuration(
-        hass,
-        result,
-        {CONF_SUB_PROFILE: "ambilight"},
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    _assert_default_virtual_power_entry_data(
-        CalculationStrategy.LUT,
-        result["data"],
-        {CONF_MANUFACTURER: "yeelight", CONF_MODEL: "YLDL01YL/ambilight"},
-    )
-
-
-async def test_manually_setup_from_library(
-    hass: HomeAssistant,
-    mock_entity_with_model_information: MockEntityWithModel,
-) -> None:
-    mock_entity_with_model_information(
-        "light.test",
-        "ikea",
-        "LED1545G12",
-        unique_id=DEFAULT_UNIQUE_ID,
-    )
-
-    result = await hass.config_entries.flow.async_init(
-        DOMAIN,
-        context={"source": config_entries.SOURCE_USER},
-    )
-
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        {"next_step_id": MENU_OPTION_LIBRARY},
-    )
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["step_id"] == "virtual_power"
-
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        {CONF_ENTITY_ID: "light.test"},
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["step_id"] == "library"
-
-    result = await _set_virtual_power_configuration(
-        hass,
-        result,
-        {CONF_CONFIRM_AUTODISCOVERED_MODEL: True},
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-
-
-async def test_manufacturer_listing_is_filtered_by_entity_domain(
-    hass: HomeAssistant,
-) -> None:
-    light_entity = MockLight("test", STATE_ON, DEFAULT_UNIQUE_ID)
-    await create_mock_light_entity(hass, light_entity)
-
-    result = await _goto_virtual_power_strategy_step(hass, CalculationStrategy.LUT)
-
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["step_id"] == "manufacturer"
-    data_schema: vol.Schema = result["data_schema"]
-    manufacturer_select: SelectSelector = data_schema.schema["manufacturer"]
-    manufacturer_options = manufacturer_select.config["options"]
-    assert {"value": "sonos", "label": "sonos"} not in manufacturer_options
-    assert {"value": "signify", "label": "signify"} in manufacturer_options
-
-
-async def test_manufacturer_listing_is_filtered_by_entity_domain2(
-    hass: HomeAssistant,
-) -> None:
-    result = await _goto_virtual_power_strategy_step(
-        hass,
-        CalculationStrategy.LUT,
-        {
-            CONF_ENTITY_ID: "switch.test",
-            CONF_MODE: CalculationStrategy.LUT,
-            CONF_UNIQUE_ID: DEFAULT_UNIQUE_ID,
-        },
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["step_id"] == "manufacturer"
-    data_schema: vol.Schema = result["data_schema"]
-    manufacturer_select: SelectSelector = data_schema.schema["manufacturer"]
-    manufacturer_options = manufacturer_select.config["options"]
-    assert {"value": "sonos", "label": "sonos"} not in manufacturer_options
-    assert {"value": "shelly", "label": "shelly"} in manufacturer_options
-
-
-async def test_advanced_power_configuration_can_be_set(hass: HomeAssistant) -> None:
-    result = await _goto_virtual_power_strategy_step(hass, CalculationStrategy.FIXED)
-    advanced_options = {
-        CONF_CALCULATION_ENABLED_CONDITION: "{{ is_state('vacuum.my_robot_cleaner', 'docked') }}",
-    }
-
-    result = await _set_virtual_power_configuration(
-        hass,
-        result,
-        {CONF_POWER: 20},
-        advanced_options,
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-
-    _assert_default_virtual_power_entry_data(
-        CalculationStrategy.FIXED,
-        result["data"],
-        {
-            CONF_FIXED: {CONF_POWER: 20},
-            CONF_CALCULATION_ENABLED_CONDITION: "{{ is_state('vacuum.my_robot_cleaner', 'docked') }}",
-        },
-    )
-
-
-async def test_daily_energy_mandatory_fields_not_supplied(hass: HomeAssistant) -> None:
-    result = await _select_sensor_type(hass, SensorType.DAILY_ENERGY)
-
-    user_input = {CONF_NAME: "My daily energy sensor"}
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input,
-    )
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["errors"]
-    assert result["errors"] == {"base": "daily_energy_mandatory"}
-
-
-async def test_create_daily_energy_entry(hass: HomeAssistant) -> None:
-    result = await _select_sensor_type(hass, SensorType.DAILY_ENERGY)
-
-    user_input = {
-        CONF_NAME: "My daily energy sensor",
-        CONF_UNIQUE_ID: DEFAULT_UNIQUE_ID,
-        CONF_VALUE: 0.5,
-        CONF_UNIT_OF_MEASUREMENT: UnitOfPower.WATT,
-    }
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input,
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    assert result["data"] == {
-        CONF_SENSOR_TYPE: SensorType.DAILY_ENERGY,
-        CONF_NAME: "My daily energy sensor",
-        CONF_DAILY_FIXED_ENERGY: {
-            CONF_UPDATE_FREQUENCY: 1800,
-            CONF_VALUE: 0.5,
-            CONF_UNIT_OF_MEASUREMENT: UnitOfPower.WATT,
-        },
-        CONF_UNIQUE_ID: DEFAULT_UNIQUE_ID,
-    }
-
-    await hass.async_block_till_done()
-    assert hass.states.get("sensor.my_daily_energy_sensor_energy")
-
-
-async def test_create_group_entry(hass: HomeAssistant) -> None:
-    result = await _select_sensor_type(hass, SensorType.GROUP)
-    user_input = {
-        CONF_NAME: "My group sensor",
-        CONF_UNIQUE_ID: DEFAULT_UNIQUE_ID,
-        CONF_GROUP_POWER_ENTITIES: ["sensor.balcony_power", "sensor.bedroom1_power"],
-    }
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input,
-    )
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    assert result["data"] == {
-        CONF_SENSOR_TYPE: SensorType.GROUP,
-        CONF_NAME: "My group sensor",
-        CONF_HIDE_MEMBERS: False,
-        CONF_GROUP_POWER_ENTITIES: ["sensor.balcony_power", "sensor.bedroom1_power"],
-        CONF_UNIQUE_ID: DEFAULT_UNIQUE_ID,
-        CONF_CREATE_UTILITY_METERS: False,
-    }
-
-    await hass.async_block_till_done()
-    assert hass.states.get("sensor.my_group_sensor_power")
-
-
-async def test_create_group_entry_without_unique_id(hass: HomeAssistant) -> None:
-    result = await _select_sensor_type(hass, SensorType.GROUP)
-    user_input = {
-        CONF_NAME: "My group sensor",
-        CONF_GROUP_POWER_ENTITIES: ["sensor.balcony_power"],
-    }
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input,
-    )
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    assert result["data"] == {
-        CONF_SENSOR_TYPE: SensorType.GROUP,
-        CONF_NAME: "My group sensor",
-        CONF_HIDE_MEMBERS: False,
-        CONF_GROUP_POWER_ENTITIES: ["sensor.balcony_power"],
-        CONF_UNIQUE_ID: "My group sensor",
-        CONF_CREATE_UTILITY_METERS: False,
-    }
-
-    await hass.async_block_till_done()
-    assert hass.states.get("sensor.my_group_sensor_power")
-
-
-async def test_group_include_area(
-    hass: HomeAssistant,
-    entity_reg: EntityRegistry,
-    area_reg: AreaRegistry,
-) -> None:
-    # Create light entity and add to group My area
-    light = MockLight("test")
-    await create_mock_light_entity(hass, light)
-    area = area_reg.async_get_or_create("My area")
-    entity_reg.async_update_entity(light.entity_id, area_id=area.id)
-
-    result = await _goto_virtual_power_strategy_step(
-        hass,
-        CalculationStrategy.FIXED,
-        {CONF_ENTITY_ID: "light.test"},
-    )
-    await _set_virtual_power_configuration(
-        hass,
-        result,
-        {CONF_STATES_POWER: {"playing": 1.8}},
-    )
-
-    result = await _select_sensor_type(hass, SensorType.GROUP)
-    user_input = {
-        CONF_NAME: "My group sensor",
-        CONF_AREA: area.id,
-        CONF_CREATE_UTILITY_METERS: True,
-    }
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input,
-    )
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    assert result["data"] == {
-        CONF_SENSOR_TYPE: SensorType.GROUP,
-        CONF_NAME: "My group sensor",
-        CONF_HIDE_MEMBERS: False,
-        CONF_AREA: area.id,
-        CONF_UNIQUE_ID: "My group sensor",
-        CONF_CREATE_UTILITY_METERS: True,
-    }
-    await hass.async_block_till_done()
-
-    power_state = hass.states.get("sensor.my_group_sensor_power")
-    assert power_state
-    assert power_state.attributes.get(CONF_ENTITIES) == {"sensor.test_power"}
-
-    energy_state = hass.states.get("sensor.my_group_sensor_energy")
-    assert energy_state
-    assert energy_state.attributes.get(CONF_ENTITIES) == {"sensor.test_energy"}
-
-    assert hass.states.get("sensor.my_group_sensor_energy_daily")
-
-
-async def test_can_unset_area(hass: HomeAssistant, area_reg: AreaRegistry) -> None:
-    area_reg.async_get_or_create("My area")
-    config_entry = MockConfigEntry(
-        domain=DOMAIN,
-        unique_id="abcdefg",
-        data={
-            CONF_SENSOR_TYPE: SensorType.GROUP,
-            CONF_AREA: "My area",
-        },
-        title="TestArea",
-    )
-    config_entry.add_to_hass(hass)
-
-    updated_entry = hass.config_entries.async_get_entry(config_entry.entry_id)
-    assert updated_entry.data == {
-        CONF_SENSOR_TYPE: SensorType.GROUP,
-        CONF_AREA: "My area",
-    }
-
-    result = await _initialize_options_flow(hass, config_entry)
-
-    user_input = {}
-    await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input=user_input,
-    )
-    updated_entry = hass.config_entries.async_get_entry(config_entry.entry_id)
-    assert updated_entry.data == {
-        CONF_SENSOR_TYPE: SensorType.GROUP,
-        CONF_CREATE_UTILITY_METERS: False,
-        CONF_HIDE_MEMBERS: False,
-    }
-
-
-async def test_can_select_existing_powercalc_entry_as_group_member(
-    hass: HomeAssistant,
-) -> None:
-    """
-    Test if we can select previously created virtual power config entries as the group member.
-    Only entries with a unique ID must be selectable
-    """
-
-    config_entry_1 = await create_mocked_virtual_power_sensor_entry(
-        hass,
-        "VirtualPower1",
-        "abcdef",
-    )
-    config_entry_2 = await create_mocked_virtual_power_sensor_entry(
-        hass,
-        "VirtualPower2",
-        None,
-    )
-    config_entry_3 = MockConfigEntry(
-        domain=DOMAIN,
-        unique_id="abcdefg",
-        data={
-            CONF_SENSOR_TYPE: SensorType.VIRTUAL_POWER,
-            CONF_UNIQUE_ID: "abcdefg",
-            CONF_ENTITY_ID: "sensor.dummy",
-            CONF_MODE: CalculationStrategy.FIXED,
-            CONF_FIXED: {CONF_POWER: 50},
-        },
-        title="VirtualPower3",
-    )
-    config_entry_3.add_to_hass(hass)
-    assert await hass.config_entries.async_setup(config_entry_3.entry_id)
-    await hass.async_block_till_done()
-
-    result = await _select_sensor_type(hass, SensorType.GROUP)
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    data_schema: vol.Schema = result["data_schema"]
-    select: SelectSelector = data_schema.schema[CONF_GROUP_MEMBER_SENSORS]
-    options = select.config["options"]
-    assert {"value": config_entry_1.entry_id, "label": "VirtualPower1"} in options
-    assert {"value": config_entry_2.entry_id, "label": "VirtualPower2"} not in options
-
-    user_input = {
-        CONF_NAME: "My group sensor",
-        CONF_UNIQUE_ID: DEFAULT_UNIQUE_ID,
-        CONF_GROUP_MEMBER_SENSORS: [config_entry_1.entry_id],
-    }
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input,
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    assert result["data"] == {
-        CONF_SENSOR_TYPE: SensorType.GROUP,
-        CONF_NAME: "My group sensor",
-        CONF_HIDE_MEMBERS: False,
-        CONF_GROUP_MEMBER_SENSORS: [config_entry_1.entry_id],
-        CONF_UNIQUE_ID: DEFAULT_UNIQUE_ID,
-        CONF_CREATE_UTILITY_METERS: False,
-    }
-
-
-async def test_group_error_mandatory(hass: HomeAssistant) -> None:
-    result = await _select_sensor_type(hass, SensorType.GROUP)
-    user_input = {CONF_NAME: "My group sensor", CONF_UNIQUE_ID: DEFAULT_UNIQUE_ID}
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input,
-    )
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["errors"]
-    assert result["errors"]["base"] == "group_mandatory"
-
-
-async def test_fixed_options_flow(hass: HomeAssistant) -> None:
-    entry = _create_mock_entry(
-        hass,
-        {
-            CONF_ENTITY_ID: "light.test",
-            CONF_SENSOR_TYPE: SensorType.VIRTUAL_POWER,
-            CONF_MODE: CalculationStrategy.FIXED,
-            CONF_FIXED: {CONF_POWER: 40},
-        },
-    )
-
-    result = await _initialize_options_flow(hass, entry)
-
-    user_input = {CONF_POWER: 50}
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input=user_input,
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    assert entry.data[CONF_FIXED][CONF_POWER] == 50
-
-
-async def test_wled_options_flow(hass: HomeAssistant) -> None:
-    await _create_wled_entities(hass)
-
-    entry = _create_mock_entry(
-        hass,
-        {
-            CONF_ENTITY_ID: "light.test",
-            CONF_SENSOR_TYPE: SensorType.VIRTUAL_POWER,
-            CONF_MODE: CalculationStrategy.WLED,
-            CONF_WLED: {CONF_VOLTAGE: 5},
-        },
-    )
-
-    result = await _initialize_options_flow(hass, entry)
-
-    user_input = {CONF_VOLTAGE: 12}
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input=user_input,
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    assert entry.data[CONF_WLED][CONF_VOLTAGE] == 12
-
-
-async def test_linear_options_flow(hass: HomeAssistant) -> None:
-    entry = _create_mock_entry(
-        hass,
-        {
-            CONF_ENTITY_ID: "light.test",
-            CONF_SENSOR_TYPE: SensorType.VIRTUAL_POWER,
-            CONF_MODE: CalculationStrategy.LINEAR,
-            CONF_LINEAR: {CONF_MIN_POWER: 10, CONF_MAX_POWER: 40},
-        },
-    )
-
-    result = await _initialize_options_flow(hass, entry)
-
-    user_input = {CONF_MAX_POWER: 50}
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input=user_input,
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    assert entry.data[CONF_LINEAR][CONF_MAX_POWER] == 50
-
-
-async def test_daily_energy_options_flow(hass: HomeAssistant) -> None:
-    entry = _create_mock_entry(
-        hass,
-        {
-            CONF_NAME: "My daily energy sensor",
-            CONF_SENSOR_TYPE: SensorType.DAILY_ENERGY,
-            CONF_DAILY_FIXED_ENERGY: {CONF_VALUE: 50},
-        },
-    )
-
-    result = await _initialize_options_flow(hass, entry)
-
-    user_input = {CONF_VALUE: 75, CONF_UNIT_OF_MEASUREMENT: UnitOfPower.WATT}
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input=user_input,
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    assert (
-        entry.data[CONF_DAILY_FIXED_ENERGY][CONF_UNIT_OF_MEASUREMENT]
-        == UnitOfPower.WATT
-    )
-    assert entry.data[CONF_DAILY_FIXED_ENERGY][CONF_VALUE] == 75
-
-
-async def test_lut_options_flow(hass: HomeAssistant) -> None:
-    entry = _create_mock_entry(
-        hass,
-        {
-            CONF_ENTITY_ID: "light.spots_kitchen",
-            CONF_SENSOR_TYPE: SensorType.VIRTUAL_POWER,
-            CONF_MODE: CalculationStrategy.LUT,
-            CONF_CREATE_ENERGY_SENSOR: True,
-            CONF_MANUFACTURER: "signify",
-            CONF_MODEL: "LCT010",
-        },
-    )
-
-    result = await _initialize_options_flow(hass, entry)
-
-    user_input = {CONF_CREATE_ENERGY_SENSOR: False}
-
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input=user_input,
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    assert not entry.data[CONF_CREATE_ENERGY_SENSOR]
-
-
-async def test_library_options_flow_raises_error_on_non_existing_power_profile(
-    hass: HomeAssistant,
-) -> None:
-    entry = _create_mock_entry(
-        hass,
-        {
-            CONF_ENTITY_ID: "light.spots_kitchen",
-            CONF_SENSOR_TYPE: SensorType.VIRTUAL_POWER,
-            CONF_MANUFACTURER: "foo",
-            CONF_MODEL: "bar",
-        },
-    )
-
-    result = await _initialize_options_flow(hass, entry)
-
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["errors"] == {"not_supported": "Power profile could not be loaded"}
-
-
-async def test_group_options_flow(hass: HomeAssistant) -> None:
-    entry = _create_mock_entry(
-        hass,
-        {
-            CONF_NAME: "Kitchen",
-            CONF_SENSOR_TYPE: SensorType.GROUP,
-            CONF_GROUP_POWER_ENTITIES: ["sensor.fridge_power"],
-        },
-    )
-
-    result = await _initialize_options_flow(hass, entry)
-
-    new_entities = ["sensor.fridge_power", "sensor.kitchen_lights_power"]
-    user_input = {CONF_GROUP_POWER_ENTITIES: new_entities}
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input=user_input,
-    )
-
-    await hass.async_block_till_done()
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    assert entry.data[CONF_GROUP_POWER_ENTITIES] == new_entities
-
-    # assert hass.states.get("sensor.kitchen_power").attributes.get(ATTR_ENTITIES) == new_entities
-
-
-async def test_linear_options_flow_error(hass: HomeAssistant) -> None:
-    entry = _create_mock_entry(
-        hass,
-        {
-            CONF_ENTITY_ID: "light.test",
-            CONF_SENSOR_TYPE: SensorType.VIRTUAL_POWER,
-            CONF_MODE: CalculationStrategy.LINEAR,
-            CONF_LINEAR: {CONF_MIN_POWER: 40, CONF_MAX_POWER: 50},
-        },
-    )
-
-    result = await _initialize_options_flow(hass, entry)
-
-    user_input = {CONF_MIN_POWER: 55, CONF_MAX_POWER: 50}
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input=user_input,
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["errors"]
-    assert result["errors"]["base"] == "linear_min_higher_as_max"
-
-
-async def test_strategy_raises_unknown_error(hass: HomeAssistant) -> None:
-    with patch(
-        "custom_components.powercalc.strategy.fixed.FixedStrategy.validate_config",
-        side_effect=StrategyConfigurationError("test"),
-    ):
-        result = await _goto_virtual_power_strategy_step(
-            hass,
-            CalculationStrategy.FIXED,
-        )
-
-        result = await hass.config_entries.flow.async_configure(
-            result["flow_id"],
-            {CONF_POWER: 20},
-        )
-
-        assert result["errors"]
-        assert result["errors"]["base"] == "unknown"
-        assert result["type"] == data_entry_flow.FlowResultType.FORM
-
-
-async def test_autodiscovered_option_flow(hass: HomeAssistant) -> None:
-    """
-    Test that we can open an option flow for an auto discovered config entry
-    """
-    entry = _create_mock_entry(
-        hass,
-        {
-            CONF_ENTITY_ID: "light.test",
-            CONF_NAME: "Test",
-            CONF_SENSOR_TYPE: SensorType.VIRTUAL_POWER,
-            CONF_MANUFACTURER: "signify",
-            CONF_MODEL: "LCT010",
-        },
-        config_entries.SOURCE_INTEGRATION_DISCOVERY,
-    )
-
-    result = await _initialize_options_flow(hass, entry)
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-
-    user_input = {CONF_CREATE_ENERGY_SENSOR: False}
-
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input=user_input,
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.CREATE_ENTRY
-    assert not entry.data[CONF_CREATE_ENERGY_SENSOR]
-
-
-async def test_subgroup_selector(hass: HomeAssistant) -> None:
-    # Create two existing group config entries
-    group1_entry = _create_mock_entry(
-        hass,
-        {
-            CONF_NAME: "Group1",
-            CONF_SENSOR_TYPE: SensorType.GROUP,
-        },
-    )
-    group2_entry = _create_mock_entry(
-        hass,
-        {
-            CONF_NAME: "Group2",
-            CONF_SENSOR_TYPE: SensorType.GROUP,
-        },
-    )
-
-    # Initialize a new config flow
-    result = await _select_sensor_type(hass, SensorType.GROUP)
-
-    # Assert the two existing groups can be selected as subgroup
-    data_schema: vol.Schema = result["data_schema"]
-    sub_group_selector: SelectSelector = data_schema.schema[CONF_SUB_GROUPS]
-    options = sub_group_selector.config["options"]
-    assert options == [
-        {"label": "Group1", "value": group1_entry.entry_id},
-        {"label": "Group2", "value": group2_entry.entry_id},
-    ]
-
-    # Create the new group
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        {
-            CONF_NAME: "Group3",
-            CONF_SUB_GROUPS: [group1_entry.entry_id, group2_entry.entry_id],
-        },
-    )
-
-    # Initialize the options flow for the newly created group
-    new_entry: ConfigEntry = result["result"]
-    result = await hass.config_entries.options.async_init(
-        new_entry.entry_id,
-        data=None,
-    )
-
-    # Assert that the group itself is not selectable as subgroup
-    data_schema: vol.Schema = result["data_schema"]
-    sub_group_selector: SelectSelector = data_schema.schema[CONF_SUB_GROUPS]
-    options = sub_group_selector.config["options"]
-    assert options == [
-        {"label": "Group1", "value": group1_entry.entry_id},
-        {"label": "Group2", "value": group2_entry.entry_id},
-    ]
-
-
-def _create_mock_entry(
-    hass: HomeAssistant,
-    entry_data: ConfigType,
-    source: str = config_entries.SOURCE_USER,
-) -> MockConfigEntry:
-    entry = MockConfigEntry(domain=DOMAIN, data=entry_data, source=source)
-    entry.add_to_hass(hass)
-
-    assert not entry.options
-    return entry
-
-
-async def _initialize_options_flow(
-    hass: HomeAssistant,
-    entry: config_entries.ConfigEntry,
-) -> FlowResult:
-    await hass.config_entries.async_setup(entry.entry_id)
-    await hass.async_block_till_done()
-    result = await hass.config_entries.options.async_init(
-        entry.entry_id,
-        data=None,
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["step_id"] == "init"
-    return result
-
-
-def _assert_default_virtual_power_entry_data(
-    strategy: CalculationStrategy,
-    config_entry_data: Mapping[str, Any],
-    expected_strategy_options: dict,
-) -> None:
-    assert (
-        config_entry_data
-        == {
-            CONF_ENTITY_ID: DEFAULT_ENTITY_ID,
-            CONF_SENSOR_TYPE: SensorType.VIRTUAL_POWER,
-            CONF_MODE: strategy,
-            CONF_CREATE_ENERGY_SENSOR: True,
-            CONF_CREATE_UTILITY_METERS: False,
-            CONF_NAME: "test",
-            CONF_UNIQUE_ID: DEFAULT_UNIQUE_ID,
-            CONF_ENERGY_INTEGRATION_METHOD: ENERGY_INTEGRATION_METHOD_LEFT,
-        }
-        | expected_strategy_options
-    )
-
-
-async def _goto_virtual_power_strategy_step(
-    hass: HomeAssistant,
-    strategy: CalculationStrategy,
-    user_input: dict[str, Any] | None = None,
-) -> FlowResult:
-    """
-    - Select the virtual power sensor type
-    - Select the given calculation strategy and put in default configuration options
-    """
-
-    if user_input is None:
-        user_input = {
-            CONF_ENTITY_ID: DEFAULT_ENTITY_ID,
-            CONF_MODE: strategy,
-            CONF_UNIQUE_ID: DEFAULT_UNIQUE_ID,
-        }
-
-    result = await _select_sensor_type(hass, SensorType.VIRTUAL_POWER)
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input,
-    )
-
-    # Lut has alternate flows depending on auto discovery, don't need to assert here
-    if strategy != CalculationStrategy.LUT:
-        assert result["step_id"] == strategy
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-
-    return result
-
-
-async def _set_virtual_power_configuration(
-    hass: HomeAssistant,
-    previous_result: FlowResult,
-    basic_options: dict[str, Any] | None = None,
-    advanced_options: dict[str, Any] | None = None,
-) -> FlowResult:
-    if basic_options is None:
-        basic_options = {}
-    result = await hass.config_entries.flow.async_configure(
-        previous_result["flow_id"],
-        basic_options,
-    )
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    if advanced_options is None:
-        advanced_options = {}
-    return await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        advanced_options,
-    )
-
-
-async def _select_sensor_type(
-    hass: HomeAssistant,
-    sensor_type: SensorType,
-) -> FlowResult:
-    """Select a sensor type from the menu"""
-    result = await hass.config_entries.flow.async_init(
-        DOMAIN,
-        context={"source": config_entries.SOURCE_USER},
-    )
-
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        {"next_step_id": sensor_type},
-    )
-
-    assert result["type"] == data_entry_flow.FlowResultType.FORM
-    assert result["step_id"] == sensor_type
-
-    return result
-
-
-async def _create_wled_entities(hass: HomeAssistant) -> None:
-    light_entity = MockLight("test", STATE_ON, DEFAULT_UNIQUE_ID)
-    await create_mock_light_entity(hass, light_entity)
-
-    platform: test_sensor_platform = getattr(hass.components, "test.sensor")
-    platform.init(empty=True)
-    estimated_current_entity = platform.MockSensor(
-        name="test_estimated_current",
-        native_value="5.0",
-        unique_id=DEFAULT_UNIQUE_ID,
-    )
-    platform.ENTITIES[0] = estimated_current_entity
-
-    assert await async_setup_component(
-        hass,
-        sensor.DOMAIN,
-        {sensor.DOMAIN: {CONF_PLATFORM: "test"}},
-    )
-    await hass.async_block_till_done()
